pyproject.toml
[project]
name = "reciepe-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.129.0",
    "motor>=3.7.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
